+++ src/role_policy/routers.py
@@ -19,6 +19,20 @@ permission_router = APIRouter(tags=["Permission"], prefix="/ref/policy/permissio
 
 
 @permission_router.get(
+    "/",
+    response_model=list[role_policy_schemas.PermissionResponse],
+)
+async def permission_get_all(
+        current_user: auth_schemas.User = Depends(auth_services.AuthService.get_current_user)
+):
+    await auth_services.UserService.check_permission(
+        user_id=current_user.id,
+        permission_title="get-list-permissions"
+    )
+    return await role_policy_services.PermissionService.get_all()
+
+
+@permission_router.get(
     "/{permission_id}/",
     response_model=role_policy_schemas.PermissionResponse,
 )
@@ -26,19 +40,14 @@ async def permission_get(
         permission_id: int,
         current_user: auth_schemas.User = Depends(auth_services.AuthService.get_current_user)
 ):
+    await auth_services.UserService.check_permission(
+        user_id=current_user.id,
+        permission_title="read-permissions"
+    )
+
     return await role_policy_services.PermissionService.get(
         permission_id=permission_id,
     )
-
-
-@permission_router.get(
-    "/",
-    response_model=list[role_policy_schemas.PermissionResponse],
-)
-async def permission_get_all(
-        current_user: auth_schemas.User = Depends(auth_services.AuthService.get_current_user)
-):
-    return await role_policy_services.PermissionService.get_all()
 
 
 @permission_router.post(
@@ -50,6 +59,10 @@ async def permission_add(
         permission_data: role_policy_schemas.PermissionRequest,
         current_user: auth_schemas.User = Depends(auth_services.AuthService.get_current_user)
 ):
+    await auth_services.UserService.check_permission(
+        user_id=current_user.id,
+        permission_title="create-permissions"
+    )
     return await role_policy_services.PermissionService.add(
         current_user_id=current_user.id,
         permission_data=permission_data
@@ -62,6 +75,10 @@ async def permission_update(
         permission_data: role_policy_schemas.PermissionRequest,
         current_user: auth_schemas.User = Depends(auth_services.AuthService.get_current_user)
 ):
+    await auth_services.UserService.check_permission(
+        user_id=current_user.id,
+        permission_title="update-permissions"
+    )
     return await role_policy_services.PermissionService.update(
         permission_id=permission_id,
         permission_data=permission_data
@@ -73,23 +90,47 @@ async def permission_delete(
         permission_id: int,
         current_user: auth_schemas.User = Depends(auth_services.AuthService.get_current_user)
 ):
+    await auth_services.UserService.check_permission(
+        user_id=current_user.id,
+        permission_title="delete-permissions"
+    )
     await role_policy_services.PermissionService.delete(
         permission_id=permission_id,
     )
     return {"message": "Permission deleted successfully"}
 
 
-role_router = APIRouter(tags=["Role"], prefix="/ref/policy/role")
-
-
-@role_router.get("/{role_id}/", response_model=role_policy_schemas.RoleResponse)
-async def role_get(
-        role_id: int,
+@permission_router.delete("/{permission_id}/soft/")
+async def permission_soft_delete(
+        permission_id: int,
         current_user: auth_schemas.User = Depends(auth_services.AuthService.get_current_user)
 ):
-    return await role_policy_services.RoleService.get(
-        role_id=role_id,
+    await auth_services.UserService.check_permission(
+        user_id=current_user.id,
+        permission_title="delete-permissions"
     )
+    await role_policy_services.PermissionService.soft_delete(
+        permission_id=permission_id,
+        current_user_id=current_user.id
+    )
+    return {"message": "Permission soft deleted successfully"}
+
+
+@permission_router.post("/{permission_id}/restore/")
+async def permission_restore(
+        permission_id: int,
+        current_user: auth_schemas.User = Depends(auth_services.AuthService.get_current_user)
+):
+    await auth_services.UserService.check_permission(
+        user_id=current_user.id,
+        permission_title="restore-permissions"
+    )
+    return await role_policy_services.PermissionService.restore(
+        permission_id=permission_id,
+    )
+
+
+role_router = APIRouter(tags=["Role"], prefix="/ref/policy/role")
 
 
 @role_router.get(
@@ -99,7 +140,26 @@ async def role_get(
 async def role_get_all(
         current_user: auth_schemas.User = Depends(auth_services.AuthService.get_current_user)
 ):
+    await auth_services.UserService.check_permission(
+        user_id=current_user.id,
+        permission_title="get-list-roles"
+    )
     return await role_policy_services.RoleService.get_all()
+
+
+@role_router.get("/{role_id}/", response_model=role_policy_schemas.RoleResponse)
+async def role_get(
+        role_id: int,
+        current_user: auth_schemas.User = Depends(auth_services.AuthService.get_current_user)
+):
+    await auth_services.UserService.check_permission(
+        user_id=current_user.id,
+        permission_title="read-roles"
+    )
+    return await role_policy_services.RoleService.get(
+        role_id=role_id,
+    )
+
 
 
 @role_router.post(
@@ -111,6 +171,10 @@ async def role_add(
         role_data: role_policy_schemas.RoleRequest,
         current_user: auth_schemas.User = Depends(auth_services.AuthService.get_current_user)
 ):
+    await auth_services.UserService.check_permission(
+        user_id=current_user.id,
+        permission_title="create-roles"
+    )
     return await role_policy_services.RoleService.add(
         role_data=role_data,
         current_user_id=current_user.id,
@@ -126,6 +190,10 @@ async def role_update(
         role_data: role_policy_schemas.RoleRequest,
         current_user: auth_schemas.User = Depends(auth_services.AuthService.get_current_user)
 ):
+    await auth_services.UserService.check_permission(
+        user_id=current_user.id,
+        permission_title="update-roles"
+    )
     return await role_policy_services.RoleService.update(
         role_id=role_id,
         role_data=role_data,
@@ -137,33 +205,45 @@ async def role_delete(
         role_id: int,
         current_user: auth_schemas.User = Depends(auth_services.AuthService.get_current_user)
 ):
+    await auth_services.UserService.check_permission(
+        user_id=current_user.id,
+        permission_title="delete-roles"
+    )
     await role_policy_services.RoleService.delete(
         role_id=role_id,
     )
     return {"message": "Role deleted successfully"}
 
 
-@role_router.delete("/{role_id}/soft")
+@role_router.delete("/{role_id}/soft/")
 async def role_soft_delete(
         role_id: int,
         current_user: auth_schemas.User = Depends(auth_services.AuthService.get_current_user)
 ):
+    await auth_services.UserService.check_permission(
+        user_id=current_user.id,
+        permission_title="delete-roles"
+    )
     await role_policy_services.RoleService.soft_delete(
         current_user_id=current_user.id,
         role_id=role_id,
     )
-    return {"message": "Role soft_deleted successfully"}
+    return {"message": "Role soft deleted successfully"}
 
 
-@role_router.delete(
-    "/{role_id}/refresh",
+@role_router.post(
+    "/{role_id}/restore/",
     response_model=role_policy_schemas.RoleResponse,
 )
-async def role_refresh(
+async def role_restore(
         role_id: int,
         current_user: auth_schemas.User = Depends(auth_services.AuthService.get_current_user)
 ):
-    return await role_policy_services.RoleService.refresh(
+    await auth_services.UserService.check_permission(
+        user_id=current_user.id,
+        permission_title="restore-roles"
+    )
+    return await role_policy_services.RoleService.restore(
         role_id=role_id,
     )
 
@@ -176,6 +256,10 @@ async def role_get_all_permission(
         role_id: int,
         current_user: auth_schemas.User = Depends(auth_services.AuthService.get_current_user)
 ):
+    await auth_services.UserService.check_permission(
+        user_id=current_user.id,
+        permission_title="read-roles"
+    )
     return await role_policy_services.RoleService.get_all_permission(
         role_id=role_id,
     )
@@ -190,6 +274,10 @@ async def role_add_permission(
         permission_id: int,
         current_user: auth_schemas.User = Depends(auth_services.AuthService.get_current_user)
 ):
+    await auth_services.UserService.check_permission(
+        user_id=current_user.id,
+        permission_title="create-roles"
+    )
     return await role_policy_services.RoleService.add_permission(
         role_id=role_id,
         permission_id=permission_id,
@@ -206,6 +294,10 @@ async def role_delete_permission(
         permission_id: int,
         current_user: auth_schemas.User = Depends(auth_services.AuthService.get_current_user)
 ):
+    await auth_services.UserService.check_permission(
+        user_id=current_user.id,
+        permission_title="delete-roles"
+    )
     return await role_policy_services.RoleService.delete_permission(
         role_id=role_id,
         permission_id=permission_id,
